log package refactored
