+++ src/main/java/com/configcat/ConfigCatClient.java
@@ -1,5 +1,8 @@
 package com.configcat;
 
+import com.configcat.log.ConfigCatLogMessages;
+import com.configcat.log.ConfigCatLogger;
+import com.configcat.log.LogLevel;
 import com.google.gson.JsonElement;
 
 import java9.util.function.Consumer;
@@ -459,34 +462,6 @@ public final class ConfigCatClient implements ConfigurationProvider {
         }
     }
 
-    private String getVariationIdFromSettingsMap(SettingResult settingResult, String key, User user, String defaultVariationId) {
-        User userObject = user != null ? user : this.defaultUser;
-        try {
-            Map<String, Setting> settings = settingResult.settings();
-            if (settings.isEmpty()) {
-                String error = ConfigCatLogMessages.getConfigJsonIsNotPresentedWithDefaultValue(key, "defaultVariationId", defaultVariationId);
-                this.hooks.invokeOnFlagEvaluated(EvaluationDetails.fromError(key, null, error, userObject));
-                this.logger.error(1000, error);
-                return defaultVariationId;
-            }
-
-            Setting setting = settings.get(key);
-            if (setting == null) {
-                String error = ConfigCatLogMessages.getSettingEvaluationFailedDueToMissingKey(key, "defaultVariationId", defaultVariationId, settings.keySet());
-                this.hooks.invokeOnFlagEvaluated(EvaluationDetails.fromError(key, null, error, userObject));
-                this.logger.error(1001, error);
-                return defaultVariationId;
-            }
-
-            return this.rolloutEvaluator.evaluate(setting, key, userObject).variationId;
-        } catch (Exception e) {
-            String error = ConfigCatLogMessages.getSettingEvaluationFailedForOtherReason(key, "defaultVariationId", defaultVariationId);
-            this.hooks.invokeOnFlagEvaluated(EvaluationDetails.fromError(key, null, error + " " + e.getMessage(), userObject));
-            this.logger.error(2001, error, e);
-            return defaultVariationId;
-        }
-    }
-
     private <T> Map.Entry<String, T> getKeyAndValueFromSettingsMap(Class<T> classOfT, SettingResult settingsResult, String variationId) {
         try {
             if (!checkSettingsAvailable(settingsResult, "null")) {
@@ -699,8 +674,9 @@ public final class ConfigCatClient implements ConfigurationProvider {
         }
 
         /**
-         * Default: Global. Set this parameter to be in sync with the Data Governance preference on the Dashboard:
-         * https://app.configcat.com/organization/data-governance (Only Organization Admins have access)
+         * Default: Global. Set this parameter to be in sync with the Data Governance preference on the Dashboard. (Only Organization Admins have access)
+         * 
+         * @see <a href="https://app.configcat.com/organization/data-governance">Data Governance</a>
          *
          * @param dataGovernance the {@link DataGovernance} parameter.
          */
