+++ src/main/java/com/configcat/log/LogLevel.java
@@ -1,4 +1,4 @@
-package com.configcat;
+package com.configcat.log;
 
 /**
  * Describes the log level.
