+++ src/test/java/com/configcat/ConfigCatClientIntegrationTest.java
@@ -1,5 +1,6 @@
 package com.configcat;
 
+import com.configcat.log.LogLevel;
 import okhttp3.OkHttpClient;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
